+++ resources/js/components/AttributeTemplate.vue
@@ -15,7 +15,14 @@
                     :key-fn="getConceptLabel"
                     :default-value="state.searchResetValue"
                     @selected="label => labelSelected(label, 'label')"
-                />
+                >
+                    <template #chain="{ option }">
+                        <ChainList
+                            :lists="option.parent_path"
+                            :max-length="3"
+                        />
+                    </template>
+                </simple-search>
             </div>
         </div>
         <div class="mb-3">
@@ -304,7 +311,12 @@
         multiselectResetClasslist,
     } from '@/helpers/helpers.js';
 
+    import ChainList from './chain/ChainList.vue';
+
     export default {
+        components: {
+            ChainList,
+        },
         props: {
             type: {
                 required: false,
@@ -524,10 +536,6 @@
                             state.attribute.rootAttributeLabel.id > 0
                         );
                 }),
-            });
-
-            // ON MOUNTED
-            onMounted(_ => {
             });
 
             // RETURN